+++ src-programs/midimixup.cpp
@@ -109,7 +109,7 @@ void randomizeNotes(vector<MidiEvent*>& notes) {
    random_device rd;
    mt19937 md(rd());
    uniform_real_distribution<double> dist(0, 100);
-   int count = notes.size();
+   int count = (int)notes.size();
    vector<pairing> neworder;
    neworder.resize(notes.size());
    int i;
